Refactor IssuePrintFormatter with new presentationIssue struct
Based on feedback from @williammartin to my question about using
interfaces for mocking in testing, I've created a new presentationIssue
struct that, when created, handles all the string formatting of an issue
that the IssuePrintFormatter needs. That removed the need for mocking
functions, as the IssuePrintFormatter now just uses any defined
presentationIssue struct.

There's still a few questions I have about this that I expect to discuss
with @williammartin and address before this is completed